requirements.txt
numpy
scipy
pandas
scikit-learn
matplotlib
tqdm
rich
pytorch-lightning
hydra-core
omegaconf
wandb
einops
opt_einsum
cmake 
transformers 
torchvision
timm
prettytable
numerize
git-lfs
torchtext
datasets 
pyfaidx
polars
genomic-benchmarks
loguru
liftover
pyyaml
